pyproject.toml
[build-system]
requires = ["setuptools >= 61"]
build-backend = "setuptools.build_meta"

[project]
name = "l10n-tools"
version = "0.0.1"
description = "Tools for localization"
readme = "README.md"
license = { text = "MPL-2.0" }
authors = [
  { name = "Mozilla", email = "l10n-drivers@mozilla.org" },
  { name = "Eemeli Aro", email = "eemeli@mozilla.com" },
  { name = "Axel Hecht" },
]
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: Mozilla Public License 2.0 (MPL 2.0)",
  "Operating System :: OS Independent",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Topic :: Software Development :: Localization",
  "Topic :: Software Development :: Testing",
]
requires-python = "~= 3.10"
dependencies = ["fluent.syntax ~= 0.19.0", "translate-toolkit ~= 3.12.2"]

[project.urls]
repository = "https://github.com/eemeli/l10n-tools"

[tool.isort]
profile = "black"
extra_standard_library = ["_typeshed"]

[tool.mypy]
exclude = "^tests/"
strict = true

[[tool.mypy.overrides]]
module = "translate.storage.properties"
ignore_missing_imports = true

[tool.setuptools]
platforms = ["any"]
packages = ["l10n_parser"]
